+++ R/helpers.R
@@ -175,3 +175,72 @@ generate_caption <- function(attributes) {
   # Return the generated caption
   return(caption)
 }
+
+#' Augment Cumulative Sum
+#'
+#' @family Utility Functions
+#' @author Steven P. Sanderson II, MPH
+#' @description This function augments a data frame by adding cumulative sum
+#' columns for specified variables.
+#'
+#' @details The function takes a data frame and a column name (or names) and
+#' computes the cumulative sum for each specified column, starting from an
+#' initial value. If the column names are not provided, it will throw an error.
+#'
+#' @param .data A data frame to augment.
+#' @param .value A column name or names for which to compute the cumulative sum.
+#' @param .names Optional. A character vector of names for the new cumulative
+#' sum columns. Defaults to "auto", which generates names based on the original
+#' column names.
+#' @param .initial_value A numeric value to start the cumulative sum from.
+#' Defaults to 0.
+#'
+#' @return A tibble with the original data and additional columns containing
+#' the cumulative sums.
+#'
+#' @examples
+#' library(dplyr)
+#' df <- tibble(x = 1:5, y = 6:10)
+#' std_cum_sum_augment(df, .value = x)
+#' std_cum_sum_augment(df, .value = y, .names = c("cumsum_y"))
+#'
+#' @name std_cum_sum_augment
+NULL
+#' @rdname std_cum_sum_augment
+#' @export
+
+std_cum_sum_augment <- function(.data,
+                                .value,
+                                .names = "auto",
+                                .initial_value = 0) {
+  column_expr <- rlang::enquo(.value)
+
+  if (rlang::quo_is_missing(column_expr)) {
+    rlang::abort("std_cum_sum_augment(.value) is missing.", use_cli_format = TRUE)
+  }
+
+  col_nms <- names(tidyselect::eval_select(column_expr, .data))
+
+  make_call <- function(col) {
+    rlang::expr(!!.initial_value + cumsum(!!rlang::sym(col)))
+  }
+
+  grid <- expand.grid(
+    col = col_nms,
+    stringsAsFactors = FALSE
+  )
+
+  calls <- purrr::pmap(list(grid$col), make_call)
+
+  if (any(.names == "auto")) {
+    newname <- paste0("cum_sum_", grid$col)
+  } else {
+    newname <- as.list(.names)
+  }
+
+  calls <- purrr::set_names(calls, newname)
+
+  ret <- tibble::as_tibble(dplyr::mutate(.data, !!!calls))
+
+  return(ret)
+}