+++ NEWS.md
@@ -1,5 +1,16 @@
 # RandomWalker (development version)
 
+## Breaking Changes
+None
+
+## New Features
+1. Fix #92 - Add function `std_cum_sum_augment()` to calculate the cumulative 
+sum of a random walk
+2 Fix 
+
+## Minor Improvements and Fixes
+None
+
 # RandomWalker 0.1.0
 
 ## Breaking Changes